Fix fq construction
daterange_fq was returning a string, but set_fq was now returning an
array. Reconciled to use an array (which solr then effectively
interprets as being ANDed together).

Another date range test was failing because of deletes now showing up;
changed that to specify a date in the future to ensure no documents
would be returned.