Break out defaults and ResumptionToken
Added a subclass of the stock ResumptionToken
and use new methods to easily create a new
token from an old one. Code in model and
partial_result now call the same method instead
of duplicating code, and allows us to put local-specific
stuff (e.g., that a new token has a last_str of "*")
in one spot.

Also extracted default query params and default
token params, and added tests to verify that we're getting
next tokens and different results for successive pages.