Enable handling single page of results & last page of results
For a single page of results, we can detect this relatively easily and
ensure we don't return a resumption token at all.

For the last page of partial results, this required some changes to
track the number of results we've returned so far. By default, Solr will
return an empty page of results along with the same cursorMark to
indicate that we're done. However, OAI requires that we return either an
error or at least one result, so waiting until we get that empty page of
results won't work.