Improve dublin core creator, subject output
Gets these from Solr (formatted in the same way we have them in the
facets) rather than what rubymarc is currently doing, which splits up
the individual subfields and omits everything after the first field if
e.g. 650 is repeated.

Ultimately, would be preferable to fix this in rubymarc, but this works
for now.